Add a package for SignInUp Activity & Profile
