+++ app/src/main/java/com/iaproject/miage/intelligentagenda/Profil/Profile.java
@@ -1,4 +1,4 @@
-package com.iaproject.miage.intelligentagenda;
+package com.iaproject.miage.intelligentagenda.Profil;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -9,6 +9,8 @@ import android.widget.TextView;
 
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
+import com.iaproject.miage.intelligentagenda.R;
+import com.iaproject.miage.intelligentagenda.SignInUP.LoginActivity;
 
 public class Profile extends AppCompatActivity {
