+++ app/src/main/java/com/iaproject/miage/intelligentagenda/SignInUP/ResetPasswordActivity.java
@@ -1,4 +1,4 @@
-package com.iaproject.miage.intelligentagenda;
+package com.iaproject.miage.intelligentagenda.SignInUP;
 
 
 import android.os.Bundle;
@@ -14,6 +14,7 @@ import android.widget.Toast;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
+import com.iaproject.miage.intelligentagenda.R;
 
 public class ResetPasswordActivity extends AppCompatActivity {
     public FirebaseAuth firebaseAuth;
