+++ app/src/main/res/layout/activity_main.xml
@@ -7,7 +7,7 @@
     android:paddingRight = "16dp"
     android:paddingTop = "16dp"
     android:paddingBottom = "16dp"
-    tools:context = ".MainActivity">
+    tools:context = ".SignInUP.MainActivity">
     android:textSize = "35dp" />
