+++ app/src/main/res/layout/activity_profile.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.iaproject.miage.intelligentagenda.Profile">
+    tools:context="com.iaproject.miage.intelligentagenda.Profil.Profile">
 
     <Button
         android:text="Logout"
