+++ app/src/main/res/layout/activity_reset_password.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="com.iaproject.miage.intelligentagenda.ResetPasswordActivity">
+    tools:context="com.iaproject.miage.intelligentagenda.SignInUP.ResetPasswordActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
